+++ packages/cli/src/merge/command.ts
@@ -1,6 +1,6 @@
 import yargs from 'yargs';
 import { Opts } from '../options';
-import { ensure, build } from '../util/command-builders';
+import { ensure, build, override } from '../util/command-builders';
 import * as o from '../options';
 
 export type MergeOptions = Required<
@@ -12,6 +12,7 @@ export type MergeOptions = Required<
     | 'removeUnmapped'
     | 'workflowMappings'
     | 'log'
+    | 'force'
   >
 >;
 
@@ -21,6 +22,9 @@ const options = [
   o.removeUnmapped,
   o.workflowMappings,
   o.log,
+  override(o.force, {
+    description: 'Force a merge even when workflows are incompatible',
+  }),
 ];
 
 const mergeCommand: yargs.CommandModule = {