+++ packages/cli/src/merge/handler.ts
@@ -51,10 +51,10 @@ const mergeHandler = async (options: MergeOptions, logger: Logger) => {
     return;
   }
 
-  // TODO pick options from the terminal
   const final = Project.merge(sourceProject, targetProject, {
     removeUnmapped: options.removeUnmapped,
     workflowMappings: options.workflowMappings,
+    force: options.force,
   });
   const yaml = final.serialize('state', { format: 'yaml' });
   await fs.writeFile(finalPath, yaml);
@@ -65,6 +65,7 @@ const mergeHandler = async (options: MergeOptions, logger: Logger) => {
       command: 'checkout',
       projectPath: commandPath,
       projectName: final.name || '',
+      log: options.log,
     },
     logger
   );